Se corrige tamaño de las variables para la matriz.
Se acomoda el llamado a lector_de_matriz_t dentro del procesador de petri.

+++ procesador_petri.h
@@ -7,14 +7,16 @@
 #define TRANSICIONES 8
 #endif
 
+#include "lector_de_matriz.h"
+
 typedef struct procesador_petri_t procesador_petri_t;
 
 struct procesador_petri_t
 {
     int transiciones;
     int plazas;
-    int matriz_estado[PLAZAS];
-    int matriz_incidencia[PLAZAS][TRANSICIONES];
+    long int matriz_estado[PLAZAS];
+    long int matriz_incidencia[PLAZAS][TRANSICIONES];
     int (*solicitud_disparo)(procesador_petri_t *petri, int t_disparar);
     int (*disparar)(procesador_petri_t *petri, int disparo);
 };
